chore: add category validation to add product

+++ src/main/java/com/val/project/repository/CategoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.val.project.entity.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Double> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
 }
